requirements.txt

jupyter==1.0.0
jupyterlab==4.0.9
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyarrow==14.0.1
python-dotenv==1.0.0
numpy==1.24.3
boto3==1.40.15